bot/requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
attrs==24.3.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
yarl==1.18.3
python-dotenv==1.0.1
SQLAlchemy==2.0.37